Unify the image tests
