+++ src/Core/tests/DeviceTests/Handlers/Image/ImageHandlerTests.cs
@@ -4,9 +4,14 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Maui.DeviceTests.Stubs;
-using Microsoft.Maui.Graphics;
-using Microsoft.Maui.Handlers;
 using Xunit;
+
+#if ANDROID
+using Android.Graphics.Drawables;
+using PlatformImageType = System.Int32;
+#elif IOS || MACCATALYST
+using PlatformImageType = UIKit.UIImage;
+#endif
 
 namespace Microsoft.Maui.DeviceTests
 {
@@ -19,6 +24,12 @@ namespace Microsoft.Maui.DeviceTests
 		where TImageHandler : IImageHandler, new()
 		where TStub : StubBase, IImageStub, new()
 	{
+#if ANDROID
+		const string ImageEventMemberName = "SetImageResource";
+#elif IOS || MACCATALYST
+		const string ImageEventMemberName = "Image";
+#endif
+
 		[Theory(
 #if IOS
 			Skip = "Test failing on IOS"
@@ -300,6 +311,132 @@ namespace Microsoft.Maui.DeviceTests
 			return events;
 		}
 
+		[Theory]
+		[InlineData("#FF0000")]
+		[InlineData("#00FF00")]
+		[InlineData("#000000")]
+		public async Task InitializingNullSourceOnlyUpdatesNull(string colorHex)
+		{
+			var expectedColor = Color.FromArgb(colorHex);
+
+			var image = new TStub
+			{
+				Background = new SolidPaintStub(expectedColor),
+			};
+
+			await InvokeOnMainThreadAsync(async () =>
+			{
+				var handler = CreateHandler<CountedImageHandler>(image);
+
+				await image.Wait();
+
+				Assert.Empty(handler.ImageEvents);
+
+				await handler.PlatformView.AssertContainsColor(expectedColor);
+			});
+		}
+
+		[Fact]
+		public async Task InitializingSourceOnlyUpdatesImageOnce()
+		{
+			var image = new TStub
+			{
+				Background = new SolidPaintStub(Colors.Black),
+				Source = new FileImageSourceStub("red.png"),
+			};
+
+			await InvokeOnMainThreadAsync(async () =>
+			{
+				var handler = CreateHandler<CountedImageHandler>(image);
+
+				await image.Wait();
+
+				await handler.PlatformView.AssertContainsColor(Colors.Red);
+
+				Assert.Single(handler.ImageEvents);
+				Assert.Equal(ImageEventMemberName, handler.ImageEvents[0].Member);
+				var platformImage = Assert.IsType<PlatformImageType>(handler.ImageEvents[0].Value);
+
+#if ANDROID
+				Assert.Equal(GetDrawableId("red"), platformImage);
+#elif IOS || MACCATALYST
+				platformImage.AssertContainsColor(Colors.Red.ToPlatform());
+#endif
+			});
+		}
+
+		[Fact]
+		public async Task UpdatingSourceOnlyUpdatesImageOnce()
+		{
+			var image = new TStub
+			{
+				Background = new SolidPaintStub(Colors.Black),
+				Source = new FileImageSourceStub("red.png"),
+			};
+
+			await InvokeOnMainThreadAsync(async () =>
+			{
+				var handler = CreateHandler<CountedImageHandler>(image);
+
+				await image.Wait();
+
+				await handler.PlatformView.AssertContainsColor(Colors.Red);
+
+				handler.ImageEvents.Clear();
+
+				image.Source = new FileImageSourceStub("blue.png");
+				handler.UpdateValue(nameof(IImage.Source));
+
+				await image.Wait();
+
+				await handler.PlatformView.AssertContainsColor(Colors.Blue);
+
+				Assert.Single(handler.ImageEvents);
+				Assert.Equal(ImageEventMemberName, handler.ImageEvents[0].Member);
+				var platformImage = Assert.IsType<PlatformImageType>(handler.ImageEvents[0].Value);
+
+#if ANDROID
+				Assert.Equal(GetDrawableId("blue"), platformImage);
+#elif IOS || MACCATALYST
+				platformImage.AssertContainsColor(Colors.Blue.ToPlatform());
+#endif
+			});
+		}
+
+		[Fact]
+		public async Task ImageLoadSequenceIsCorrectWithChecks()
+		{
+			var events = await ImageLoadSequenceIsCorrect();
+
+			Assert.Single(events);
+			Assert.Equal(ImageEventMemberName, events[0].Member);
+
+#if ANDROID
+			var platformImage = Assert.IsType<ColorDrawable>(events[0].Value);
+			platformImage.Color.IsEquivalent(Colors.Blue.ToPlatform());
+#elif IOS || MACCATALYST
+			var platformImage = Assert.IsType<UIImage>(events[0].Value);
+			platformImage.AssertContainsColor(Colors.Blue.ToPlatform());
+#endif
+		}
+
+		[Fact]
+		public async Task InterruptingLoadCancelsAndStartsOverWithChecks()
+		{
+			var events = await InterruptingLoadCancelsAndStartsOver();
+
+			Assert.Single(events);
+			Assert.Equal(ImageEventMemberName, events[0].Member);
+
+#if ANDROID
+			var platformImage = Assert.IsType<ColorDrawable>(events[0].Value);
+			platformImage.Color.IsEquivalent(Colors.Red.ToPlatform());
+#elif IOS || MACCATALYST
+			var platformImage = Assert.IsType<UIImage>(events[0].Value);
+			platformImage.AssertContainsColor(Colors.Red.ToPlatform());
+#endif
+		}
+
 		protected TCustomHandler CreateHandler<TCustomHandler>(IView view)
 			where TCustomHandler : IImageHandler, new()
 		{
@@ -315,5 +452,10 @@ namespace Microsoft.Maui.DeviceTests
 
 			return handler;
 		}
+
+#if ANDROID
+		static int GetDrawableId(string image) =>
+			MauiProgram.DefaultContext.Resources.GetDrawableId(MauiProgram.DefaultContext.PackageName, image);
+#endif
 	}
 }
